floats:size

+++ db/migrate/20170620182854_add_unit_to_accessories.rb
@@ -1,6 +1,6 @@
 class AddUnitToAccessories < ActiveRecord::Migration[5.0]
   def change
     add_reference :accessories, :unit, foreign_key: true, index: true
-    add_column :accessories, :measure, :integer
+    add_column :accessories, :measure, :decimal
   end
 end
